Fix layout and few styling fixes
