+++ frontend/src/app/auth/popup-callback/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 
-import { useEffect, useState } from 'react';
-import { useSearchParams } from 'next/navigation';
+import { useEffect, useState } from "react";
+import { useSearchParams } from "next/navigation";
 
 /**
  * This page is specifically designed to handle OAuth popup callbacks.
@@ -9,46 +9,54 @@ import { useSearchParams } from 'next/navigation';
  * them to the parent window that opened this popup.
  */
 export default function PopupCallbackPage() {
-    const [error, setError] = useState<string | null>(null);
-    const searchParams = useSearchParams();
+  const [error, setError] = useState<string | null>(null);
+  const searchParams = useSearchParams();
 
-    useEffect(() => {
-        try {
-            const token = searchParams.get('token');
-            const uid = searchParams.get('uid');
-            const errorMsg = searchParams.get('error');
+  useEffect(() => {
+    try {
+      const token = searchParams.get("token");
+      const uid = searchParams.get("uid");
+      const errorMsg = searchParams.get("error");
 
-            // Verify window.opener exists
-            if (!window.opener) {
-                setError('Authentication window was not opened correctly');
-                return;
-            }
+      // Verify window.opener exists
+      if (!window.opener) {
+        setError("Authentication window was not opened correctly");
+        return;
+      }
 
-            // Send message to parent window
-            if (errorMsg) {
-                window.opener.postMessage({ error: decodeURIComponent(errorMsg) }, window.location.origin);
-            } else if (token && uid) {
-                window.opener.postMessage({ token, uid }, window.location.origin);
-            } else {
-                window.opener.postMessage({ error: 'Invalid authentication data received' }, window.location.origin);
-            }
+      // Send message to parent window
+      if (errorMsg) {
+        window.opener.postMessage(
+          { error: decodeURIComponent(errorMsg) },
+          window.location.origin
+        );
+      } else if (token && uid) {
+        window.opener.postMessage({ token, uid }, window.location.origin);
+      } else {
+        window.opener.postMessage(
+          { error: "Invalid authentication data received" },
+          window.location.origin
+        );
+      }
 
-            // Close the popup after sending the message
-            window.close();
-        } catch (err: any) {
-            console.error('Error in popup callback:', err);
-            setError(err instanceof Error ? err.message : 'Failed to process authentication');
-        }
-    }, [searchParams]);
+      // Close the popup after sending the message
+      window.close();
+    } catch (err: any) {
+      console.error("Error in popup callback:", err);
+      setError(
+        err instanceof Error ? err.message : "Failed to process authentication"
+      );
+    }
+  }, [searchParams]);
 
-    return (
-        <div className="popup-callback-container">
-            <h1>Authentication Complete</h1>
-            {error ? (
-                <p>Error: {error}</p>
-            ) : (
-                <p>Authentication successful! You can close this window.</p>
-            )}
-        </div>
-    );
+  return (
+    <div className='popup-callback-container'>
+      <h1>Authentication Complete</h1>
+      {error ? (
+        <p>Error: {error}</p>
+      ) : (
+        <p>Authentication successful! You can close this window.</p>
+      )}
+    </div>
+  );
 }
