+++ frontend/src/app/page.tsx
@@ -7,17 +7,11 @@ import Footer from "@/components/Footer";
 
 export default function Page() {
   return (
-    <AuthProvider>
-      <Background />
-      <Navbar />
-
-      <div className='main-content'>
-        <div className='header'>
-          <h1>Summarizz</h1>
-          <p>Summarize your articles, videos, and more with ease.</p>
-        </div>
+    <div className='main-content'>
+      <div className='header'>
+        <h1>Summarizz</h1>
+        <p>Summarize your articles, videos, and more with ease.</p>
       </div>
-      <Footer />
-    </AuthProvider>
+    </div>
   );
 }
