+++ frontend/src/app/profile/[id]/manage/page.tsx
@@ -10,7 +10,6 @@ import { User } from "@/models/User";
 
 import "@/app/styles/profile/ProfileManagement.scss";
 import { useParams } from "next/navigation";
-import Layout from "@/app/content/layout";
 
 /**
  * Page() -> JSX.Element
@@ -467,343 +466,335 @@ export default function Page() {
   // -------------- Render ----------------
   // --------------------------------------
   return (
-    <Layout>
-      <div className='main-content'>
-        {/******************** EDIT PROFILE  ********************/}
-        <div className='profile-management-section'>
-          <h2>Edit Profile</h2>
+    <div className='main-content'>
+      {/******************** EDIT PROFILE  ********************/}
+      <div className='profile-management-section'>
+        <h2>Edit Profile</h2>
 
-          <form onSubmit={handleEditProfile}>
-            {/* TODO: Profile Image */}
-            <div className='profile-image-section'>
-              <div className='input-group'>
-                {profileImagePreview ? (
-                  <Image
-                    src={profileImagePreview}
-                    width={200}
-                    height={200}
-                    alt='Profile Picture'
-                    className='profile-edit-image'
-                  />
-                ) : (
-                  <h1 className='profile-edit-image profile-initial'>
-                    {user?.username[0].toUpperCase()}
-                  </h1>
-                )}
-              </div>
-
-              <div>
-                <label htmlFor='profile-image' className='profile-image-upload'>
-                  {profileImage ? "Change" : "Upload"} Profile Image
-                </label>
-                <input
-                  id='profile-image'
-                  type='file'
-                  accept='image/*'
-                  onChange={handleProfileImageChange}
-                />
-                {imageError && <p className='error-message'>{imageError}</p>}
-              </div>
-            </div>
-
-            <div className='form-group'>
-              {/* TODO: First Name */}
-              <div className='input-group'>
-                <label htmlFor='firstName'>First Name *</label>
-                <input
-                  type='text'
-                  id='firstName'
-                  placeholder='First Name'
-                  value={user?.firstName ? user.firstName : ""}
-                  onChange={(e) =>
-                    setUser(
-                      user ? { ...user, firstName: e.target.value } : null
-                    )
-                  }
-                />
-              </div>
-
-              {/* TODO: Last Name */}
-              <div className='input-group'>
-                <label htmlFor='lastName'>Last Name *</label>
-                <input
-                  type='text'
-                  id='lastName'
-                  placeholder='Last Name'
-                  value={user?.lastName ? user.lastName : ""}
-                  onChange={(e) =>
-                    setUser(user ? { ...user, lastName: e.target.value } : null)
-                  }
-                />
-              </div>
-            </div>
-
-            {/* TODO: Bio */}
+        <form onSubmit={handleEditProfile}>
+          {/* TODO: Profile Image */}
+          <div className='profile-image-section'>
             <div className='input-group'>
-              <label htmlFor='bio'>Bio</label>
-              <textarea
-                id='bio'
-                placeholder='Tell us about yourself...'
-                value={user?.bio ? user.bio : ""}
-                onChange={(e) =>
-                  setUser(user ? { ...user, bio: e.target.value } : null)
-                }
-              />
+              {profileImagePreview ? (
+                <Image
+                  src={profileImagePreview}
+                  width={200}
+                  height={200}
+                  alt='Profile Picture'
+                  className='profile-edit-image'
+                />
+              ) : (
+                <h1 className='profile-edit-image profile-initial'>
+                  {user?.username[0].toUpperCase()}
+                </h1>
+              )}
             </div>
 
-            {/* TODO: Phone */}
-            <div className='input-group'>
-              <label htmlFor='phone'>Phone</label>
+            <div>
+              <label htmlFor='profile-image' className='profile-image-upload'>
+                {profileImage ? "Change" : "Upload"} Profile Image
+              </label>
               <input
-                type='tel'
-                id='phone'
-                placeholder='(123) 321-1234'
-                value={user?.phone ? user.phone : ""}
-                onChange={(e) =>
-                  setUser(user ? { ...user, phone: e.target.value } : null)
-                }
+                id='profile-image'
+                type='file'
+                accept='image/*'
+                onChange={handleProfileImageChange}
               />
+              {imageError && <p className='error-message'>{imageError}</p>}
             </div>
-
-            {/* TODO: Date of Birth */}
-            <div className='input-group'>
-              <label htmlFor='dob'>Date of Birth</label>
-              <input
-                type='date'
-                id='dob'
-                value={user?.dateOfBirth ? user.dateOfBirth : ""}
-                onChange={(e) =>
-                  setUser(
-                    user ? { ...user, dateOfBirth: e.target.value } : null
-                  )
-                }
-              />
-            </div>
-
-            {/* Privacy Setting */}
-            <div className="input-group">
-              <label htmlFor="isPrivate">Make Profile Private</label>
-              <input
-                type="checkbox"
-                id="isPrivate"
-                checked={user?.isPrivate || false}
-                onChange={(e) =>
-                  setUser(user ? { ...user, isPrivate: e.target.checked } : null)
-                }
-              />
-            </div>
-
-            {errorEditProfile && (
-              <p className='error-message'>{errorEditProfile}</p>
-            )}
-
-            {successEditProfile && (
-              <p className='success-message'>{successEditProfile}</p>
-            )}
-
-            <button type='submit' className='save-button'>
-              Save Changes
-            </button>
-          </form>
-        </div>
-
-        {/******************** EDIT CREDENTIALS  ********************/}
-        <div className='profile-management-section'>
-          <h2>Edit Credentials</h2>
-
-          {/* TABS */}
-          <div className='tabs'>
-            <button
-              className={
-                "tab first-tab" + (activeTab === "password" ? " active-tab" : "")
-              }
-              onClick={() => setActiveTab("password")}
-            >
-              Change Password
-            </button>
-            <button
-              className={"tab" + (activeTab === "email" ? " active-tab" : "")}
-              onClick={() => setActiveTab("email")}
-            >
-              Change Email
-            </button>
-            <button
-              className={
-                "tab last-tab" + (activeTab === "username" ? " active-tab" : "")
-              }
-              onClick={() => setActiveTab("username")}
-            >
-              Change Username
-            </button>
           </div>
 
-          {/* PASSWORD */}
-          {activeTab === "password" && (
-            <div className='tab-content'>
-              <h3>Change Password</h3>
-              <form onSubmit={handleChangePassword}>
-                <div className='input-group'>
-                  <label htmlFor='currentPassword'>Current Password</label>
-                  <input
-                    type='password'
-                    id='currentPassword'
-                    value={currentPassword}
-                    onChange={(e) => setCurrentPassword(e.target.value)}
-                    required
-                  />
-                </div>
-
-                <div className='form-group'>
-                  <div className='input-group'>
-                    <label htmlFor='newPassword'>New Password</label>
-                    <input
-                      type='password'
-                      id='newPassword'
-                      value={newPassword}
-                      onChange={(e) => setNewPassword(e.target.value)}
-                    />
-                  </div>
-
-                  <div className='input-group'>
-                    <label htmlFor='confirmPassword'>
-                      Confirm New Password
-                    </label>
-                    <input
-                      type='password'
-                      id='confirmPassword'
-                      value={confirmPassword}
-                      onChange={(e) => setConfirmPassword(e.target.value)}
-                    />
-                  </div>
-                </div>
-
-                {errorEditPassword && (
-                  <p className='error-message'>{errorEditPassword}</p>
-                )}
-                {successEditPassord && (
-                  <p className='success-message'>{successEditPassord}</p>
-                )}
-
-                <button type='submit' className='save-button'>
-                  Change Password
-                </button>
-              </form>
+          <div className='form-group'>
+            {/* TODO: First Name */}
+            <div className='input-group'>
+              <label htmlFor='firstName'>First Name *</label>
+              <input
+                type='text'
+                id='firstName'
+                placeholder='First Name'
+                value={user?.firstName ? user.firstName : ""}
+                onChange={(e) =>
+                  setUser(user ? { ...user, firstName: e.target.value } : null)
+                }
+              />
             </div>
-          )}
 
-          {/* EMAIL */}
-
-          {activeTab === "email" && (
-            <div className='tab-content'>
-              <h3>Change Email</h3>
-              <form onSubmit={handleUpdateEmail}>
-                <div className='input-group'>
-                  <label htmlFor='newEmail'>New Email</label>
-                  <input
-                    type='email'
-                    id='newEmail'
-                    value={newEmail}
-                    onChange={(e) => setNewEmail(e.target.value)}
-                  />
-                </div>
-
-                <div className='input-group'>
-                  <label htmlFor='currentPassword'>Enter Password</label>
-                  <input
-                    type='password'
-                    id='currentPassword'
-                    value={currentPassword}
-                    onChange={(e) => setCurrentPassword(e.target.value)}
-                    required
-                  />
-                </div>
-
-                {errorEditEmail && (
-                  <p className='error-message'>{errorEditEmail}</p>
-                )}
-                {successEditEmail && (
-                  <p className='success-message'>{successEditEmail}</p>
-                )}
-
-                <button type='submit' className='save-button'>
-                  Change Email
-                </button>
-              </form>
+            {/* TODO: Last Name */}
+            <div className='input-group'>
+              <label htmlFor='lastName'>Last Name *</label>
+              <input
+                type='text'
+                id='lastName'
+                placeholder='Last Name'
+                value={user?.lastName ? user.lastName : ""}
+                onChange={(e) =>
+                  setUser(user ? { ...user, lastName: e.target.value } : null)
+                }
+              />
             </div>
-          )}
+          </div>
 
-          {/* USERNAME */}
-          {activeTab === "username" && (
-            <div className='tab-content'>
-              <h3>Change Username</h3>
-              <form onSubmit={handleUpdateUsername}>
-                <div className='input-group'>
-                  <label htmlFor='newUsername'>New Username</label>
-                  <input
-                    type='text'
-                    id='newUsername'
-                    value={newUsername}
-                    onChange={(e) => setNewUsername(e.target.value)}
-                  />
-                </div>
-
-                {errorEditUsername && (
-                  <p className='error-message'>{errorEditUsername}</p>
-                )}
-                {successEditUsername && (
-                  <p className='success-message'>{successEditUsername}</p>
-                )}
-
-                <button type='submit' className='save-button'>
-                  Change Username
-                </button>
-              </form>
-            </div>
-          )}
-        </div>
-
-        {/******************** DELETE ACCOUNT  ********************/}
-        <div className='profile-management-section'>
-          <h2>Delete Account</h2>
-
+          {/* TODO: Bio */}
           <div className='input-group'>
-            <label htmlFor='email'>Enter Email</label>
-            <input
-              type='email'
-              id='email'
-              value={email}
-              onChange={(e) => setEmail(e.target.value)}
+            <label htmlFor='bio'>Bio</label>
+            <textarea
+              id='bio'
+              placeholder='Tell us about yourself...'
+              value={user?.bio ? user.bio : ""}
+              onChange={(e) =>
+                setUser(user ? { ...user, bio: e.target.value } : null)
+              }
             />
           </div>
 
+          {/* TODO: Phone */}
           <div className='input-group'>
-            <label htmlFor='password'>Enter Password</label>
+            <label htmlFor='phone'>Phone</label>
             <input
-              type='password'
-              id='password'
-              value={password}
-              onChange={(e) => setPassword(e.target.value)}
+              type='tel'
+              id='phone'
+              placeholder='(123) 321-1234'
+              value={user?.phone ? user.phone : ""}
+              onChange={(e) =>
+                setUser(user ? { ...user, phone: e.target.value } : null)
+              }
             />
           </div>
 
-          <p>
-            By pressing "Delete Account", I acknowledge the account will be
-            deleted from this platform as well as all the content related. There
-            is no recovery once an account is deleted.
-          </p>
+          {/* TODO: Date of Birth */}
+          <div className='input-group'>
+            <label htmlFor='dob'>Date of Birth</label>
+            <input
+              type='date'
+              id='dob'
+              value={user?.dateOfBirth ? user.dateOfBirth : ""}
+              onChange={(e) =>
+                setUser(user ? { ...user, dateOfBirth: e.target.value } : null)
+              }
+            />
+          </div>
 
-          {errorDeleteAccount && (
-            <p className='error-message'>{errorDeleteAccount}</p>
-          )}
-          {successDeleteAccount && (
-            <p className='success-message'>{successDeleteAccount}</p>
+          {/* Privacy Setting */}
+          <div className='input-group'>
+            <label htmlFor='isPrivate'>Make Profile Private</label>
+            <input
+              type='checkbox'
+              id='isPrivate'
+              checked={user?.isPrivate || false}
+              onChange={(e) =>
+                setUser(user ? { ...user, isPrivate: e.target.checked } : null)
+              }
+            />
+          </div>
+
+          {errorEditProfile && (
+            <p className='error-message'>{errorEditProfile}</p>
           )}
 
-          <button onClick={handleDeleteAccount} className='save-button warning'>
-            Delete Account
+          {successEditProfile && (
+            <p className='success-message'>{successEditProfile}</p>
+          )}
+
+          <button type='submit' className='save-button'>
+            Save Changes
+          </button>
+        </form>
+      </div>
+
+      {/******************** EDIT CREDENTIALS  ********************/}
+      <div className='profile-management-section'>
+        <h2>Edit Credentials</h2>
+
+        {/* TABS */}
+        <div className='tabs'>
+          <button
+            className={
+              "tab first-tab" + (activeTab === "password" ? " active-tab" : "")
+            }
+            onClick={() => setActiveTab("password")}
+          >
+            Change Password
+          </button>
+          <button
+            className={"tab" + (activeTab === "email" ? " active-tab" : "")}
+            onClick={() => setActiveTab("email")}
+          >
+            Change Email
+          </button>
+          <button
+            className={
+              "tab last-tab" + (activeTab === "username" ? " active-tab" : "")
+            }
+            onClick={() => setActiveTab("username")}
+          >
+            Change Username
           </button>
         </div>
+
+        {/* PASSWORD */}
+        {activeTab === "password" && (
+          <div className='tab-content'>
+            <h3>Change Password</h3>
+            <form onSubmit={handleChangePassword}>
+              <div className='input-group'>
+                <label htmlFor='currentPassword'>Current Password</label>
+                <input
+                  type='password'
+                  id='currentPassword'
+                  value={currentPassword}
+                  onChange={(e) => setCurrentPassword(e.target.value)}
+                  required
+                />
+              </div>
+
+              <div className='form-group'>
+                <div className='input-group'>
+                  <label htmlFor='newPassword'>New Password</label>
+                  <input
+                    type='password'
+                    id='newPassword'
+                    value={newPassword}
+                    onChange={(e) => setNewPassword(e.target.value)}
+                  />
+                </div>
+
+                <div className='input-group'>
+                  <label htmlFor='confirmPassword'>Confirm New Password</label>
+                  <input
+                    type='password'
+                    id='confirmPassword'
+                    value={confirmPassword}
+                    onChange={(e) => setConfirmPassword(e.target.value)}
+                  />
+                </div>
+              </div>
+
+              {errorEditPassword && (
+                <p className='error-message'>{errorEditPassword}</p>
+              )}
+              {successEditPassord && (
+                <p className='success-message'>{successEditPassord}</p>
+              )}
+
+              <button type='submit' className='save-button'>
+                Change Password
+              </button>
+            </form>
+          </div>
+        )}
+
+        {/* EMAIL */}
+
+        {activeTab === "email" && (
+          <div className='tab-content'>
+            <h3>Change Email</h3>
+            <form onSubmit={handleUpdateEmail}>
+              <div className='input-group'>
+                <label htmlFor='newEmail'>New Email</label>
+                <input
+                  type='email'
+                  id='newEmail'
+                  value={newEmail}
+                  onChange={(e) => setNewEmail(e.target.value)}
+                />
+              </div>
+
+              <div className='input-group'>
+                <label htmlFor='currentPassword'>Enter Password</label>
+                <input
+                  type='password'
+                  id='currentPassword'
+                  value={currentPassword}
+                  onChange={(e) => setCurrentPassword(e.target.value)}
+                  required
+                />
+              </div>
+
+              {errorEditEmail && (
+                <p className='error-message'>{errorEditEmail}</p>
+              )}
+              {successEditEmail && (
+                <p className='success-message'>{successEditEmail}</p>
+              )}
+
+              <button type='submit' className='save-button'>
+                Change Email
+              </button>
+            </form>
+          </div>
+        )}
+
+        {/* USERNAME */}
+        {activeTab === "username" && (
+          <div className='tab-content'>
+            <h3>Change Username</h3>
+            <form onSubmit={handleUpdateUsername}>
+              <div className='input-group'>
+                <label htmlFor='newUsername'>New Username</label>
+                <input
+                  type='text'
+                  id='newUsername'
+                  value={newUsername}
+                  onChange={(e) => setNewUsername(e.target.value)}
+                />
+              </div>
+
+              {errorEditUsername && (
+                <p className='error-message'>{errorEditUsername}</p>
+              )}
+              {successEditUsername && (
+                <p className='success-message'>{successEditUsername}</p>
+              )}
+
+              <button type='submit' className='save-button'>
+                Change Username
+              </button>
+            </form>
+          </div>
+        )}
       </div>
-    </Layout>
+
+      {/******************** DELETE ACCOUNT  ********************/}
+      <div className='profile-management-section'>
+        <h2>Delete Account</h2>
+
+        <div className='input-group'>
+          <label htmlFor='email'>Enter Email</label>
+          <input
+            type='email'
+            id='email'
+            value={email}
+            onChange={(e) => setEmail(e.target.value)}
+          />
+        </div>
+
+        <div className='input-group'>
+          <label htmlFor='password'>Enter Password</label>
+          <input
+            type='password'
+            id='password'
+            value={password}
+            onChange={(e) => setPassword(e.target.value)}
+          />
+        </div>
+
+        <p>
+          By pressing "Delete Account", I acknowledge the account will be
+          deleted from this platform as well as all the content related. There
+          is no recovery once an account is deleted.
+        </p>
+
+        {errorDeleteAccount && (
+          <p className='error-message'>{errorDeleteAccount}</p>
+        )}
+        {successDeleteAccount && (
+          <p className='success-message'>{successDeleteAccount}</p>
+        )}
+
+        <button onClick={handleDeleteAccount} className='save-button warning'>
+          Delete Account
+        </button>
+      </div>
+    </div>
   );
 }
