+++ frontend/src/app/styles/content/toolbar.scss
@@ -1,4 +1,4 @@
-select {
+.toolbar select {
   width: 200px;
   padding: 5px;
   box-sizing: border-box;
@@ -64,7 +64,12 @@ select {
     line-height: 1.6;
     color: #333;
 
-    h1, h2, h3, h4, h5, h6 {
+    h1,
+    h2,
+    h3,
+    h4,
+    h5,
+    h6 {
       margin: 16px 0;
       font-weight: bold;
     }
@@ -73,7 +78,8 @@ select {
       margin: 12px 0;
     }
 
-    ul, ol {
+    ul,
+    ol {
       margin: 12px 0;
       padding-left: 20px;
     }
@@ -93,4 +99,3 @@ select {
     }
   }
 }
-
