+++ frontend/src/app/styles/footer.scss
@@ -1,18 +1,14 @@
 .footer {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    margin-top: auto;
-    box-sizing: border-box;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  margin-top: auto;
+  box-sizing: border-box;
+  margin: 20px 0px;
 }
 
 .footer-logo {
-    width: 50px;
-    height: 50px;
-    margin: 20 0 5 0; 
-    order: 1;
-}
-
-p {
-    order: 2;
+  width: 50px;
+  height: 50px;
+  margin: 20 0 5 0;
 }
