+++ frontend/src/components/NavbarWrapper.tsx
@@ -1,0 +1,13 @@
+"use client";
+
+import { usePathname } from "next/navigation";
+import Navbar from "./Navbar";
+
+export default function NavbarWrapper() {
+  const pathname = usePathname();
+  const showNavbar = !pathname.startsWith("/auth");
+
+  if (!showNavbar) return null;
+
+  return <Navbar />;
+}
